Implement display of filtered movies in FiltersList.vue
This update modifies the applyFilter function within FiltersList.vue to store the response data in a newly defined filteredMovies reference. It then adds a section in the component's template to list out these filtered movies. Therefore, users now can see the movies list based on the selected filters directly on the application.