Extend FilterModal with fieldType dependent inputs
Extended the functionality of the FilterModal by introducing inputs that are dependent on the type of fields. This change includes the addition of a new fieldType property in the criteria object, and conditions to show either DatePicker, InputNumber, or InputText based on the fieldType. This allows for better input validation.