Update default value for criteriaValue in FilterModal.vue
Changed the default criteriaValue in FilterModal.vue from an empty string to null. This adjustment was made in several spots such as the constants declaration, the addCriteria function, and the criteria value reset in order to better handle data types that are not strings. Also enhanced the input number field attributes for better usability.